refactoring main files
