+++ common/adb.py
